Merge commits up to v5.0.2:
* Start WiX Toolset v5.0.2

Re-enable insignificant zeros in PackageVersion

Closes 8589

Fix crash with use of File/@PatchGroup.

Fix crash when adding file in patch when filtering

Add overloads to support create-only Wixouts.

This prevents the .NET ZipArchive (and friends) from keeping the whole
thing in memory, to support updating when we don't need to update the
Wixout when building a binary Wixlib.

WiX Toolset v5.0.2